First attempt of the new version (works quite well)
Remove PLBEvents + Silke for mode Manual

+++ test code/PLB/ESP industry/test/MockBuilding.h
@@ -9,7 +9,7 @@
 class MockBuilding : public IBuilding
 {
 public:
-    MOCK_METHOD(int, calculateSolarPower, ());
+    MOCK_METHOD(float, calculateSolarPower, ());
     MOCK_METHOD(void, charge, (float));
 };
